requirements.txt
pyyaml
toml
openai
